benerin

+++ database/migrations/2025_05_08_034418_create_certificates_table.php
@@ -7,7 +7,8 @@ use Illuminate\Support\Facades\Schema;
 class CreateCertificatesTable extends Migration
 {
     public function up()
-    {
+{
+    if (!Schema::hasTable('certificates')) {
         Schema::create('certificates', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
@@ -18,6 +19,8 @@ class CreateCertificatesTable extends Migration
             $table->timestamps();
         });
     }
+}
+
 
     public function down()
     {
